requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
aiosqlite==0.21.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.5
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
discord.py==2.6.3
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lyricsgenius==3.7.2
multidict==6.4.3
propcache==0.3.1
pycparser==2.22
PyNaCl==1.6.0
python-dotenv==1.1.0
requests==2.32.4
sniffio==1.3.1
soupsieve==2.8
spotipy==2.25.1
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.0
yt-dlp==2025.10.22 #asw